requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
dataclasses==0.8
decorator==5.0.9
docopt==0.6.2
docutils==0.17.1
Flask==2.0.1
Flask-HTTPAuth==4.4.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
html2text==2020.1.16
idna==2.10
idna-ssl==1.1.0
importlib-metadata==4.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
libsass==0.12.3
lxml==4.6.3
MarkupSafe==2.0.1
meraki==1.10.0
multidict==5.1.0
num2words==0.5.10
numpy==1.19.4
passlib==1.7.4
Pillow==8.3.1
polib==1.1.1
psutil==5.8.0
psycopg2==2.9.1
PyPDF2==1.26.0
python-dateutil==2.8.1
python-stdnum==1.16
pytz==2021.1
PyYAML==5.4.1
reportlab==3.5.68
requests==2.25.1
schedule==1.1.0
six==1.16.0
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.8
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.2
yarl==1.6.3
zipp==3.4.1
